requirements.txt
elasticsearch==5.4.0
python-daemon==2.1.2
PyYAML==3.12
